+++ google-books-application/src/app/modules/books/components/books-favorite/books-favorite.component.ts
@@ -4,6 +4,10 @@ import { LoaderService } from '../../../../shared/services/loader.service';
 import { finalize } from 'rxjs';
 import { takeUntilDestroyed } from '@angular/core/rxjs-interop';
 import { BookItem } from '../../models/book-item.interface';
+import { NotificationService } from '../../../../shared/services/notification.service';
+import { FAVORITE_OPERATION_MESSAGE } from '../../constants/favorite-operation-message.const';
+import { DEFAULT_MESSAGE } from '../../../../shared/constants/default-message.const';
+import { ICON_DEFINITION } from '../../../../shared/constants/icon-definition.const';
 @Component({
   selector: 'app-books-favorite',
   templateUrl: './books-favorite.component.html',
@@ -11,11 +15,13 @@ import { BookItem } from '../../models/book-item.interface';
 })
 export class BooksFavoriteComponent implements OnInit {
   protected books: BookItem[] | null = null;
+  protected readonly ICON_DEFINITION = ICON_DEFINITION;
   private readonly destroyRef: DestroyRef = inject(DestroyRef);
 
   constructor(
     private booksService: BooksService,
-    private loaderService: LoaderService
+    private loaderService: LoaderService,
+    private notificationService: NotificationService
   ) {}
 
   ngOnInit(): void {
@@ -38,16 +44,42 @@ export class BooksFavoriteComponent implements OnInit {
             return;
           }
           this.books = response;
-          console.log(this.books, 'books');
+        },
+        error: () => {
+          this.notificationService.showSnackbar(DEFAULT_MESSAGE.ERROR);
         },
       });
   }
 
-  protected onRemoveFromFavoriteById(id: string): void {
+  protected onClearBookById(id: string): void {
     if (!this.books) {
       return;
     }
     const bookIndex = this.books.findIndex((book) => book.id === id);
     this.books.splice(bookIndex, 1);
+  }
+
+  protected onClearAllBooks(): void {
+    if (!this.books?.length) {
+      return;
+    }
+    this.loaderService.start();
+    this.booksService
+      .clearAllBooksFromFavoriteShelf()
+      .pipe(
+        finalize(() => this.loaderService.stop()),
+        takeUntilDestroyed(this.destroyRef)
+      )
+      .subscribe({
+        next: () => {
+          this.notificationService.showSnackbar(
+            FAVORITE_OPERATION_MESSAGE.REMOVED_ALL
+          );
+          this.fetchFavoriteBooks();
+        },
+        error: () => {
+          this.notificationService.showSnackbar(DEFAULT_MESSAGE.ERROR);
+        },
+      });
   }
 }