+++ google-books-application/src/app/modules/books/components/books-search-form/components/books-search-form.component.ts
@@ -1,6 +1,7 @@
 import { Component, Output, EventEmitter } from '@angular/core';
 import { NonNullableFormBuilder, Validators } from '@angular/forms';
 import { SearchParams } from '../../../models/search-params.interface';
+import { ICON_DEFINITION } from '../../../../../shared/constants/icon-definition.const';
 @Component({
   selector: 'app-books-search-form',
   templateUrl: './books-search-form.component.html',
@@ -10,6 +11,7 @@ export class BooksSearchFormComponent {
   @Output() searchButtonClick = new EventEmitter<SearchParams>();
 
   protected formGroup;
+  protected readonly ICON_DEFINITION = ICON_DEFINITION;
 
   constructor(private formBuilder: NonNullableFormBuilder) {
     this.formGroup = this.formBuilder.group({