+++ google-books-application/src/app/modules/books/constants/favorite-operation-message.const.ts
@@ -1,4 +1,5 @@
 export const FAVORITE_OPERATION_MESSAGE = {
-  ADDED: 'The book was successfully added to "favorite"',
-  REMOVED: 'The book was successfully removed from "favorite"',
+  ADDED: 'The book was successfully added to "favorite".',
+  REMOVED: 'The book was successfully removed from "favorite".',
+  REMOVED_ALL: 'All books were successfully deleted from "favorite".',
 };