+++ google-books-application/src/app/shared/constants/default-message.const.ts
@@ -1,0 +1,3 @@
+export const DEFAULT_MESSAGE = {
+  ERROR: 'An error was occured.',
+};