+++ google-books-application/src/app/shared/constants/icon-definition.const.ts
@@ -1,0 +1,8 @@
+export const ICON_DEFINITION = {
+  FAVORITE: 'favorite',
+  FAVORITE_OUTLINE: 'favorite_outline',
+  CLOSE: 'close',
+  SEARCH: 'search',
+  LOGOUT: 'logout',
+  ARROW_DOWN: 'south',
+};